# Conflicts:
#	.fontcustom-manifest.json
#	config/sidekiq_client.yml
#	public/assets/glyphs_3a8b8805b1f1b4d332f754ce7daafc1c.eot
#	public/assets/glyphs_3a8b8805b1f1b4d332f754ce7daafc1c.svg
#	public/assets/glyphs_3a8b8805b1f1b4d332f754ce7daafc1c.ttf
#	public/assets/glyphs_3a8b8805b1f1b4d332f754ce7daafc1c.woff
#	public/assets/glyphs_41dc35fc1f90e34a8bfabb6adeedf8c2.eot
#	public/assets/glyphs_41dc35fc1f90e34a8bfabb6adeedf8c2.svg
#	public/assets/glyphs_41dc35fc1f90e34a8bfabb6adeedf8c2.ttf
#	public/assets/glyphs_41dc35fc1f90e34a8bfabb6adeedf8c2.woff
#	public/assets/glyphs_75800162d2d38887288afa700bee24b1.svg
#	public/assets/glyphs_c60762bbccdd03d21175843b10ba12c3.eot
#	public/assets/glyphs_c60762bbccdd03d21175843b10ba12c3.ttf
#	public/assets/glyphs_c60762bbccdd03d21175843b10ba12c3.woff
#	public/assets/index.html
#	public/src/fonts/_glyphs.scss
#	spec/support/forum_helper.rb

+++ app/models/helpdesk/note/rabbitmq.rb
@@ -46,7 +46,7 @@ class Helpdesk::Note < ActiveRecord::Base
     { 
       "object"                =>  "note",
       "action"                =>  action,
-      "action_epoch"          =>  Time.zone.now.to_i,
+      "action_epoch"          =>  Time.zone.now.to_f,
       "note_properties"       =>  mq_reports_note_properties(action),
       "subscriber_properties" =>  {"reports" => mq_reports_subscriber_properties(action).merge(options) }    
     }
